Add AddDocumentLink Component
