+++ src/app/components/AddDocumentLink/AddDocumentLink.stories.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import AddDocumentLink from './AddDocumentLink';
+
+export default {
+  component: AddDocumentLink,
+  title: 'Components/AddDocumentLink',
+};
+
+export const Default = (): JSX.Element => <AddDocumentLink />;
