+++ src/app/components/AddDocumentLink/AddDocumentLink.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function AddDocumentLink(): JSX.Element {
+  return (
+    <a href="/home">
+      <img src="src/app/assets/plus.svg" />
+    </a>
+  );
+}
